Create README.md
The Smart Dustbin with Blynk 2.0 is a revolutionary waste management system designed to solve the problem of overflowing waste bins. 
It utilizes wireless sensors to detect nearby presence and monitor the trash level in the bin. When someone throws trash into the bin, the system automatically opens the lid, providing a convenient and hygienic user experience.
 Additionally, the system sends alerts to the waste management agency when the bin is full, enabling efficient waste collection. By promoting a cleaner and disease-free environment, this smart dustbin aims to improve the overall quality of life in communities.
 With its seamless connectivity through Blynk 2.0, scalability, and adaptability, the Smart Dustbin system has the potential to revolutionize waste management practices and contribute to a more sustainable future.